test: add helpful error messages when tests fail

+++ test/install/test_install_config.sh
@@ -30,11 +30,11 @@ fi
 # run the install script non-interactively
 ./helix-auth-svc/install.sh -n
 
-grep -q 'User=perforce' /etc/systemd/system/helix-auth.service
-grep -q 'Group=perforce' /etc/systemd/system/helix-auth.service
+grep -q 'User=perforce' /etc/systemd/system/helix-auth.service || { echo 'service missing User' ; exit 1; }
+grep -q 'Group=perforce' /etc/systemd/system/helix-auth.service || { echo 'service missing Group' ; exit 1; }
 
 # ensure the systemd service is running
-sudo systemctl status helix-auth | grep -q 'Active: active'
+sudo systemctl status helix-auth | grep -q 'Active: active' || { echo 'service not active' ; exit 1; }
 
 # for convenience in testing, let perforce have password-less sudo
 sudo tee /etc/sudoers.d/perforce >/dev/null <<EOT
@@ -50,16 +50,16 @@ sudo -u perforce ./helix-auth-svc/bin/configure-auth-service.sh -n \
 
 # ensure configure script created the OIDC client secret file
 test -f helix-auth-svc/client-secret.txt
-sudo -u perforce grep -q 'client_secret' helix-auth-svc/client-secret.txt
-grep -q 'https://localhost:3000' helix-auth-svc/.env
-grep -q 'https://oidc.issuer' helix-auth-svc/.env
+sudo -u perforce grep -q 'client_secret' helix-auth-svc/client-secret.txt || { echo 'client secret missing' ; exit 1; }
+grep -q 'https://localhost:3000' helix-auth-svc/.env || { echo '.env missing URL' ; exit 1; }
+grep -q 'https://oidc.issuer' helix-auth-svc/.env || { echo '.env missing OIDC' ; exit 1; }
 
 # run the configure script and set up SAML
 sudo -u perforce ./helix-auth-svc/bin/configure-auth-service.sh -n \
     --base-url https://localhost:3000 \
     --saml-idp-metadata-url https://saml.idp/metadata
 
-grep -q 'https://saml.idp/metadata' helix-auth-svc/.env
+grep -q 'https://saml.idp/metadata' helix-auth-svc/.env || { echo '.env missing SAML' ; exit 1; }
 
 cat <<EOT
